Table-replication-of-Dan-Kermit-and-Seth-2002-
This work is an assignment for ECOM90003 Applied Micro-econometrics at the Univery of Melbourne. This work provides do files that replicate some statistical results in the paper "the Impact of Economic Conditions on Participation in Disability Programs: Evidence from the Coal Boom and Bust". Table 2, 3, 6, 7, 8 and 9 are included in this work.